+++ docs/changelog.md
@@ -6,6 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-01-15
+
+### Added
+
+ - Add config options `tcp-retry-count`, `tcp-retry-redispatch`, `tcp-load-balancing-algorithm`, 
+  `tcp-load-balancing-consistent-hashing`.
+ - Fix bug where integrating with `haproxy-route-tcp` does not update the relation data.
+
+## 2026-01-12
+
+### Added
+
+ - Add external-grpc-port configuration to support gRPC traffic routing.
+
+## 2025-11-25
+
+### Added
+
+ - Added upgrade documentation.
+
 ## 2025-08-01
 
 ### Changed
@@ -21,24 +41,24 @@ Each revision is versioned by the date of the revision.
 
 ## 2025-07-14
 
-### Added 
+### Added
 
 - Add check-interval, check-rise, check-fall, check-path and check-port configurations
 
 
 ## 2025-07-10
 
-### Added 
+### Added
 
 - Add retry-count, retry-interval and retry-redispatch configurations
 
-### Added 
+### Added
 
 - Add ingress integration support
 
 
 ## 2025-07-09
 
-### Added 
+### Added
 
 - Add path and subdomain configuration to the `ingress-configurator` charm.